proxylib-go: move pkg/kafka to proxylib/kafka/kafkalib
Package `pkg/kafka` provides a kafka library implementation
that is only used by the kafka parser.

To reduce the surface of the "API" of the module `cilium/proxy`,
this commit moves the package into `proxylib/kafka/kafkalib`.
