--- pkg/kafka/doc.go
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-// Package kafka provides a library to parse Kafka requests and responses and
-// apply policy rules
-package kafka
