+++ proxylib/kafka/kafkalib/doc.go
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+// Package kafkalib provides a library to parse Kafka requests and responses and
+// apply policy rules
+package kafkalib
