+++ proxylib/kafka/kafkalib/policy.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package kafka
+package kafkalib
 
 import (
 	"github.com/sirupsen/logrus"
