+++ proxylib/kafka/kafkalib/response.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package kafka
+package kafkalib
 
 import (
 	"encoding/binary"
